Environment and Enemy
Environment was imported, lighting was edited using environmental lighting to make the scene look darker and more horror like. Flashlight mechanic was added using spotlight, and attaching it to the player object. I also wrote a script that turned the flashlight on with F. AN enemy was added from the asset store.