Implementing Flashlight Health + Enemy AI
Added a 30 second timer to the flashlight, and when 30 seconds is done the flashlight runs out and displays a dialogue message on the screen. A slider was also implemented to show the flashlight running out of power. Enemy AI implementation was started, the enemy can walk forwards backwards left and right, triggering the animations as they do so, and they make their way to waypoints around the map.