Updated flashlight logic
Now batteries need to be collected in order to replenish the flashlight. I imported a battery prefab and assigned a large box collider to it with a trigger so when the player walks in the collider the "press [E] to collect" message is displayed and allows the player to pick up the battery. Once the battery is collected the flashlight can be used for another 30 seconds and a "battery collected" message is shown.