Update to enemy AI + Deleted unnecessary assets
To make the enemy AI more challenging and realistic by having it continue to search for the player in the vicinity of their last known location before resuming its patrol route.

Tracking Last Known Position: Introduced a private variable lastKnownPlayerPosition to store the player's last known location when the enemy loses sight of them.

Search Behaviour:

When the enemy loses sight of the player (canSeePlayer and canSeePlayerFromBehind are both false), it switches to a search mode (isSearchingForPlayer = true) and sets its destination to lastKnownPlayerPosition.
This behavior is triggered in the Update() method when isChasingPlayer is set to false.
Returning to Patrol:

Added a check in the Update() method to determine if the enemy has reached the last known position of the player.
If the player is not found upon reaching this position (determined by a distance check), isSearchingForPlayer is set to false, and the enemy resumes its patrol by calling the Patrol() method.